requirements.txt
python-dotenv
requests
vk-api